+++ btcstaking-tracker/stakingeventwatcher/stakingeventwatcher_test.go
@@ -1,7 +1,9 @@
 package stakingeventwatcher
 
 import (
+	"context"
 	"github.com/babylonlabs-io/babylon/testutil/datagen"
+	btcstakingtypes "github.com/babylonlabs-io/babylon/x/btcstaking/types"
 	"github.com/babylonlabs-io/vigilante/btcclient"
 	"github.com/babylonlabs-io/vigilante/config"
 	"github.com/babylonlabs-io/vigilante/metrics"
@@ -61,6 +63,7 @@ func TestHandlingDelegations(t *testing.T) {
 			DelegationStartHeight: 0,
 			UnbondingOutput:       nil,
 			HasProof:              false,
+			Status:                btcstakingtypes.BTCDelegationStatus_VERIFIED.String(),
 		})
 	}
 
@@ -89,10 +92,114 @@ func TestHandlingDelegations(t *testing.T) {
 	}
 	mockBTCClient.EXPECT().TxDetails(gomock.Any(), gomock.Any()).Return(details, btcclient.TxInChain, nil).AnyTimes()
 
-	sew.wg.Add(2)
+	sew.wg.Add(3)
 	go func() {
 		go sew.fetchDelegations()
 		go sew.handlerVerifiedDelegations()
+		go sew.handleUnbondedDelegations()
+	}()
+
+	require.Eventually(t, func() bool {
+		return promtestutil.ToFloat64(sew.metrics.ReportedActivateDelegationsCounter) >= float64(expectedActivated)
+	}, 60*time.Second, 100*time.Millisecond)
+}
+
+func TestHandlingDelegationsByEvents(t *testing.T) {
+	t.Parallel()
+	r := rand.New(rand.NewSource(time.Now().Unix()))
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+
+	cfg := config.DefaultBTCStakingTrackerConfig()
+	cfg.CheckDelegationsInterval = 1 * time.Second
+	cfg.FetchCometBlockInterval = 1 * time.Second
+
+	mockBTCClient := mocks.NewMockBTCClient(ctrl)
+	mockBabylonNodeAdapter := NewMockBabylonNodeAdapter(ctrl)
+
+	mockBabylonNodeAdapter.EXPECT().BtcClientTipHeight().Return(uint32(0), nil).AnyTimes()
+	bsMetrics := metrics.NewBTCStakingTrackerMetrics()
+
+	sew := StakingEventWatcher{
+		logger:                          zap.NewNop().Sugar(),
+		quit:                            make(chan struct{}),
+		cfg:                             &cfg,
+		babylonNodeAdapter:              mockBabylonNodeAdapter,
+		btcClient:                       mockBTCClient,
+		unbondingTracker:                NewTrackedDelegations(),
+		pendingTracker:                  NewTrackedDelegations(),
+		verifiedInsufficientConfTracker: NewTrackedDelegations(),
+		verifiedNotInChainTracker:       NewTrackedDelegations(),
+		verifiedSufficientConfTracker:   NewTrackedDelegations(),
+		unbondingDelegationChan:         make(chan *newDelegation),
+		unbondingRemovalChan:            make(chan *delegationInactive),
+		activationLimiter:               semaphore.NewWeighted(30),
+		metrics:                         bsMetrics.UnbondingWatcherMetrics,
+	}
+
+	sew.delegationRetrievalInProgress.Store(false)
+
+	mockBabylonNodeAdapter.EXPECT().CometBFTTipHeight(gomock.Any()).DoAndReturn(func(_ context.Context) (int64, error) {
+		return sew.currentCometTipHeight.Load() + 1, nil
+	}).AnyTimes()
+
+	defer close(sew.quit)
+
+	expectedActivated := 1000
+	stakingTxHashes := make([]string, 0, expectedActivated)
+
+	for i := 0; i < expectedActivated; i++ {
+		stk := datagen.GenRandomTx(r)
+		delegation := Delegation{
+			StakingTx:             stk,
+			StakingOutputIdx:      0,
+			DelegationStartHeight: 0,
+			UnbondingOutput:       nil,
+			HasProof:              false,
+			Status:                btcstakingtypes.BTCDelegationStatus_VERIFIED.String(),
+		}
+		stkHash := stk.TxHash()
+		stakingTxHashes = append(stakingTxHashes, stkHash.String())
+		mockBabylonNodeAdapter.EXPECT().
+			BTCDelegation(stkHash.String()).
+			Return(&delegation, nil).AnyTimes()
+	}
+
+	firstCall := mockBabylonNodeAdapter.EXPECT().
+		StakingTxHashesByEvent(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
+		Return(stakingTxHashes, nil).Times(1)
+
+	mockBabylonNodeAdapter.EXPECT().
+		StakingTxHashesByEvent(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
+		After(firstCall).
+		Return(nil, nil).AnyTimes()
+
+	mockBabylonNodeAdapter.EXPECT().
+		ActivateDelegation(gomock.Any(), gomock.Any(), gomock.Any()).
+		Return(nil).AnyTimes()
+
+	mockBabylonNodeAdapter.EXPECT().QueryHeaderDepth(gomock.Any()).Return(uint32(2), nil).AnyTimes()
+	mockBabylonNodeAdapter.EXPECT().IsDelegationVerified(gomock.Any()).Return(true, nil).AnyTimes()
+
+	params := BabylonParams{ConfirmationTimeBlocks: 1}
+	mockBabylonNodeAdapter.EXPECT().Params().Return(&params, nil).AnyTimes()
+
+	block, _ := datagen.GenRandomBtcdBlock(r, 10, nil)
+	bh := block.BlockHash()
+	details := &chainntnfs.TxConfirmation{
+		BlockHash:   &bh,
+		BlockHeight: 100,
+		TxIndex:     1,
+		Tx:          nil,
+		Block:       block,
+	}
+	mockBTCClient.EXPECT().TxDetails(gomock.Any(), gomock.Any()).Return(details, btcclient.TxInChain, nil).AnyTimes()
+
+	sew.wg.Add(3)
+	go func() {
+		go sew.fetchCometBftBlockForever()
+		go sew.handlerVerifiedDelegations()
+		go sew.handleUnbondedDelegations()
 	}()
 
 	require.Eventually(t, func() bool {